refactor next auth config
