+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,53 +1,5 @@
-import dbConnect from "@/lib/db";
-import { isMatchingPassword } from "@/lib/password";
-import User from "@/models/User";
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-
-export const authOptions: AuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        username: { label: "email", type: "text", placeholder: "jsmith" },
-        password: { label: "password", type: "password" },
-      },
-      async authorize(credentials, req) {
-        if (!credentials) {
-          return null;
-        }
-
-        const { username, password } = credentials;
-
-        await dbConnect();
-        const existingUser = await User.findOne({ email: username });
-        if (!existingUser) {
-          console.error("no user found for given email");
-          return null;
-        }
-
-        const isCorrectPassword = await isMatchingPassword(
-          password,
-          existingUser.passwordHash
-        );
-
-        if (!isCorrectPassword) {
-          console.error("invalid credentials");
-          return null;
-        }
-
-        return existingUser;
-      },
-    }),
-  ],
-  session: {
-    maxAge: 6 * 60 * 60,
-  },
-  pages: {
-    signIn: "/login",
-    newUser: "/signup",
-  },
-};
+import NextAuth from "next-auth";
+import { authOptions } from "./config";
 
 const handler = NextAuth(authOptions);
