+++ src/app/login/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../api/auth/[...nextauth]/config";
 import LoginUI from "./ui";
 import { redirect } from "next/navigation";
