+++ src/app/middleware/auth.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../api/auth/[...nextauth]/config";
 
 interface AuthCallbackFunc {
   (email: string): Promise<Response>;
